# openapi.yaml
openapi: 3.1.0
info:
  title: Survey Platform API
  version: 0.1.0
  description: |
    Admin REST API for the Survey Platform — used by the admin webapp and
    by the survey-platform CLI / MCP server.

    All endpoints require admin authentication: either a NextAuth session
    cookie (set when logging in via /admin/login) or an
    `Authorization: Bearer <ADMIN_API_KEY>` header.
  contact:
    name: Survey Platform
servers:
  - url: "{server}"
    description: Survey Platform deployment
    variables:
      server:
        default: http://localhost:3000
        description: Base URL of the Survey Platform deployment
security:
  - bearerAuth: []
...
paths:
...
  /api/courses/{courseId}/questions:
    get:
      tags: [questions]
      summary: List questions in a course
      operationId: listQuestions
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - name: topicId
          in: query
          required: false
          description: Filter to a single topic
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [questions]
      summary: Create a single question
      operationId: createQuestion
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/courses/{courseId}/questions/{questionId}:
    patch:
      tags: [questions]
      summary: Update a question, migrating existing student answers
      description: >
        Answers store the option text, not an option id, so renaming an option
        rewrites the stored answers in the same transaction. Moving the correct
        answer, removing an option or changing the question type re-grades
        earlier answers and therefore requires confirmRegrade when the question
        already has answers or practice attempts; without it the endpoint
        returns 409 with requiresConfirmation. FSRS scheduling (grade) is left
        untouched.
      operationId: updateQuestion
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateQuestionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Confirmation required before earlier answers are re-graded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  requiresConfirmation:
                    type: boolean
                  answers:
                    type: integer
                  practiceAttempts:
                    type: integer
    delete:
      tags: [questions]
      summary: Delete a question
      operationId: deleteQuestion
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: ADMIN_API_KEY configured server-side
  parameters:
...
    QuestionId:
      name: questionId
      in: path
      required: true
      description: Question ID
      schema:
        type: integer
        format: int64
        minimum: 1
    CourseId:
      name: courseId
      in: path
      required: true
      description: Course ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed for this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    QuestionType:
      type: string
      enum: [MULTIPLE_CHOICE, FREE_TEXT]
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Topic:
      type: object
      required: [id, name, courseId]
      properties:
        id:
          type: integer
        name:
          type: string
        courseId:
          type: integer
        _count:
          type: object
          properties:
            questions:
              type: integer
...
    Option:
      type: object
      required: [id, text, isCorrect]
      properties:
        id:
          type: integer
        text:
          type: string
        isCorrect:
          type: boolean
        questionId:
          type: integer
    Question:
      type: object
      required: [id, text, type, topicId]
      properties:
        id:
          type: integer
        text:
          type: string
        type:
          $ref: "#/components/schemas/QuestionType"
        topicId:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/Option"
        topic:
          $ref: "#/components/schemas/Topic"
    UpdateQuestionRequest:
      type: object
      description: At least one field besides confirmRegrade must be present
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
        type:
          $ref: "#/components/schemas/QuestionType"
        topicId:
          type: integer
          minimum: 1
          description: Must be a topic in the same course
        options:
          type: array
          maxItems: 10
          description: >
            Full option set for MULTIPLE_CHOICE. Keep an existing option by
            sending its id; omit the id to add a new one; leave an option out to
            remove it. Texts must be unique and at most one may be correct.
          items:
            type: object
            required: [text]
            properties:
              id:
                type: integer
                minimum: 1
              text:
                type: string
                minLength: 1
                maxLength: 500
              isCorrect:
                type: boolean
                default: false
        subskill:
          type: string
          nullable: true
        config:
          type: object
          description: SORTING configuration (categories and items)
        exemplars:
          type: array
          items:
            type: object
        confirmRegrade:
          type: boolean
          default: false
          description: Confirms that earlier answers may be re-graded
    UpdateQuestionResponse:
      type: object
      required: [question, impact]
      properties:
        question:
          $ref: "#/components/schemas/Question"
        impact:
          type: object
          description: What the change did to stored answers
          properties:
            renamedOptions:
              type: integer
            removedOptions:
              type: integer
            addedOptions:
              type: integer
            migratedAnswers:
              type: integer
            migratedAttempts:
              type: integer
            orphanedAnswers:
              type: integer
              description: Answers whose option no longer exists
            regradedAnswers:
              type: integer
            regradedAttempts:
              type: integer
    CreateQuestionRequest:
      type: object
      required: [text, type, topicId]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
        type:
          $ref: "#/components/schemas/QuestionType"
        topicId:
          type: integer
          minimum: 1
        options:
          type: array
          description: Option texts (MULTIPLE_CHOICE only)
          items:
            type: string
        correctOptionIndex:
          type: integer
          minimum: 0
          description: Index into options that is the correct answer
...